feat(Writers): Début de l'ajout de la vérification d'identité
[13][Done]
Effacer uniquement si on est admin, modifier uniquement si on est admin, ajouter seulement si on est connecté

+++ Backend/app/controllers/writers_controller.ts
@@ -1,3 +1,4 @@
+import User from '#models/user'
 import Writer from '#models/writer'
 import { WriterValidator } from '#validators/writer'
 import type { HttpContext } from '@adonisjs/core/http'
@@ -7,13 +8,18 @@ export default class WritersController {
    * Display a list of resource
    */
   async index({ response }: HttpContext) {
-    response.ok(await Writer.all())
+    response.ok(
+      await Writer.query().orderBy('lastname').orderBy('firstname').preload('books').paginate(1, 20)
+    )
   }
 
   /**
    * Handle form submission for the create action
    */
-  async store({ request, response }: HttpContext) {
+  async store({ auth, request, response }: HttpContext) {
+    // Make sure the user exists
+    await User.findOrFail(auth.user?.id)
+
     const { firstname, lastname } = await request.validateUsing(WriterValidator)
     response.created(await Writer.create({ firstname, lastname }))
   }
@@ -22,13 +28,26 @@ export default class WritersController {
    * Show individual record
    */
   async show({ params, response }: HttpContext) {
-    response.ok(await Writer.findOrFail(params.id))
+    response.ok(
+      await Writer.query()
+        .where('id', params.id)
+        .orderBy('lastname')
+        .orderBy('firstname')
+        .preload('books')
+        .firstOrFail()
+    )
   }
 
   /**
    * Handle form submission for the edit action
    */
-  async update({ params, request, response }: HttpContext) {
+  async update({ auth, params, request, response }: HttpContext) {
+    // Check to see if the user is an admin, otherwise return the 403 status code
+    if (auth.user?.role !== 'admin') {
+      response.forbidden({ message: 'Only admins can update writer information.' })
+      return
+    }
+
     const { firstname, lastname } = await request.validateUsing(WriterValidator)
     const writer = await Writer.findOrFail(params.id)
     writer.merge({ firstname, lastname })
@@ -38,7 +57,12 @@ export default class WritersController {
   /**
    * Delete record
    */
-  async destroy({ params, response }: HttpContext) {
+  async destroy({ auth, params, response }: HttpContext) {
+    // Check to see if the user is an admin, otherwise return the 403 status code
+    if (auth.user?.role !== 'admin') {
+      response.forbidden({ message: 'Only admins can delete writers.' })
+      return
+    }
     const writer = await Writer.findOrFail(params.id)
     await writer.delete()
     response.noContent()
